Added error reporting and improved file watch
Gulp script should now report WHY it won't uglify or browserify a file rather than give an ambiguous error. It seems that (as of this commit) that neither of these libraries support ES6. Avoid using things let and const.

Also, Gulp script should now watch for changes to anything in the dist folder